+++ p2p/net/swarm/swarm.go
@@ -100,6 +100,23 @@ func WithResourceManager(m network.ResourceManager) Option {
 	}
 }
 
+// WithNoDialDelay configures swarm to dial all addresses for a peer without
+// any delay
+func WithNoDialDelay() Option {
+	return func(s *Swarm) error {
+		s.dialRanker = noDelayRanker
+		return nil
+	}
+}
+
+// WithDialRanker configures swarm to use d as the DialRanker
+func WithDialRanker(d network.DialRanker) Option {
+	return func(s *Swarm) error {
+		s.dialRanker = d
+		return nil
+	}
+}
+
 // Swarm is a connection muxer, allowing connections to other peers to
 // be opened and closed, while still using the same Chan for all
 // communication. The Chan sends/receives Messages, which note the
@@ -163,6 +180,8 @@ type Swarm struct {
 
 	bwc           metrics.Reporter
 	metricsTracer MetricsTracer
+
+	dialRanker network.DialRanker
 }
 
 // NewSwarm constructs a Swarm.
@@ -181,6 +200,7 @@ func NewSwarm(local peer.ID, peers peerstore.Peerstore, eventBus event.Bus, opts
 		dialTimeout:      defaultDialTimeout,
 		dialTimeoutLocal: defaultDialTimeoutLocal,
 		maResolver:       madns.DefaultResolver,
+		dialRanker:       DefaultDialRanker,
 	}
 
 	s.conns.m = make(map[peer.ID][]*Conn)