+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Table Of Contents <!-- omit in toc -->
+- [v0.28.0](#v0280)
 - [v0.27.0](#v0270)
 - [v0.26.4](#v0264)
 - [v0.26.3](#v0263)
@@ -7,6 +8,17 @@
 - [v0.26.0](#v0260)
 - [v0.25.1](#v0251)
 - [v0.25.0](#v0250)
+
+# [v0.28.0]()
+
+## 🔦 Highlights <!-- omit in toc -->
+
+### Smart Dialing <!-- omit in toc -->
+* When connecting to a peer we now do [happy eyeballs](https://www.rfc-editor.org/rfc/rfc8305) like dial prioritisation to prefer QUIC addresses over TCP addresses. We dial the QUIC address first and wait 250ms to dial the TCP address of the peer.
+* In our experiments we've seen little impact on latencies up to 80th percentile. 90th and 95th percentile latencies are impacted. For details see discussion on the [PR](https://github.com/libp2p/go-libp2p/pull/2260#issuecomment-1528848170).
+* For details of the address ranking logic see godoc for `swarm.DefaultDialRanker`.
+* To disable smart dialing and keep the old behaviour use the 
+`libp2p.NoDelayNetworkDialRanker` option.
 
 # [v0.27.0](https://github.com/libp2p/go-libp2p/releases/tag/v0.27.0)